requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
model-bakery==1.20.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.9.0
pytest-responses==0.5.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
sqlparse==0.5.2
urllib3==2.2.3
